+++ adapter/deploy.go
@@ -10,8 +10,10 @@ import (
 	"time"
 
 	"github.com/gin-gonic/gin"
+	"gitlab.com/nunet/device-management-service/db"
 	"gitlab.com/nunet/device-management-service/docker"
 	"gitlab.com/nunet/device-management-service/models"
+	"gitlab.com/nunet/device-management-service/statsdb"
 	"gitlab.com/nunet/device-management-service/utils"
 	grpc "google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
@@ -51,6 +53,8 @@ func messageHandler(message string) {
 		kernelFilePath := "/etc/nunet/vmlinux"
 		filesystemURL := "https://d.nunet.io/fc/nunet-fc-ubuntu-20.04-0.ext4"
 		filesystemPath := "/etc/nunet/nunet-fc-ubuntu-20.04-0.ext4"
+		pKey := adapterMessage.Message.Params.PublicKey
+		nodeId := adapterMessage.Message.Params.NodeID
 
 		err = utils.DownloadFile(kernelFileURL, kernelFilePath)
 		if err != nil {
@@ -71,9 +75,13 @@ func messageHandler(message string) {
 		startDefaultBody := struct {
 			KernelImagePath string `json:"kernel_image_path"`
 			FilesystemPath  string `json:"filesystem_path"`
+			PublicKey       string `json:"public_key"`
+			NodeID          string `json:"node_id"`
 		}{
 			KernelImagePath: kernelFilePath,
 			FilesystemPath:  filesystemPath,
+			PublicKey:       pKey,
+			NodeID:          nodeId,
 		}
 		jsonBody, _ := json.Marshal(startDefaultBody)
 
@@ -87,6 +95,37 @@ func messageHandler(message string) {
 
 	} else if adapterMessage.Message.ServiceType == "ml-training-gpu" {
 		depResp := models.DeploymentResponse{}
+
+		callID := statsdb.GetCallID()
+		peerIDOfServiceHost := adapterMessage.Message.Params.NodeID
+		timeStamp := float32(statsdb.GetTimestamp())
+		status := "accepted"
+
+		ServiceCallParams := models.ServiceCall{
+			CallID:              callID,
+			PeerIDOfServiceHost: peerIDOfServiceHost,
+			ServiceID:           adapterMessage.Message.ServiceType,
+			CPUUsed:             float32(adapterMessage.Message.Constraints.CPU),
+			MaxRAM:              float32(adapterMessage.Message.Constraints.Vram),
+			MemoryUsed:          float32(adapterMessage.Message.Constraints.RAM),
+			NetworkBwUsed:       0.0,
+			TimeTaken:           0.0,
+			Status:              status,
+			Timestamp:           timeStamp,
+		}
+		statsdb.ServiceCall(ServiceCallParams)
+
+		requestTracker := models.RequestTracker{
+			ServiceType: adapterMessage.Message.ServiceType,
+			CallID:      callID,
+			NodeID:      peerIDOfServiceHost,
+			Status:      status,
+		}
+		result := db.DB.Create(&requestTracker)
+		if result.Error != nil {
+			panic(result.Error)
+		}
+
 		depResp = docker.HandleDeployment(adapterMessage.Message, depResp)
 		jsonDepResp, _ := json.Marshal(depResp)
 		SendMessage(adapterMessage.Sender, string(jsonDepResp))